feat: created a single Tree class
